- compare resolv.conf files only if they exist
git-svn-id: https://svn.berlios.de/svnroot/repos/kiwi/kiwi-head@1792 92bc7f66-5d17-0410-ac0d-e081219a1bb3

+++ modules/KIWIRoot.pm
@@ -849,12 +849,14 @@ sub setup {
 	#========================================
 	# cleanup temporary copy of resolv.conf
 	#----------------------------------------
-	my $data = qxx ("diff -q /etc/resolv.conf $root/etc/resolv.conf");
-	my $code = $? >> 8;
-	if ($code == 0) {
-		$kiwi -> info ("Cleanup temporary copy of resolv.conf");
-		qxx ("rm -f $root/etc/resolv.conf");
-		$kiwi -> done ();
+	if ((-f "$root/etc/resolv.conf") && (-f "/etc/resolv.conf")) {
+		my $data = qxx ("diff -q /etc/resolv.conf $root/etc/resolv.conf");
+		my $code = $? >> 8;
+		if ($code == 0) {
+			$kiwi -> info ("Cleanup temporary copy of resolv.conf");
+			qxx ("rm -f $root/etc/resolv.conf");
+			$kiwi -> done ();
+		}
 	}
 	#========================================
 	# cleanup temporary .buildenv
